Create bot interface implementation
