+++ services/service_creator.py
@@ -14,9 +14,9 @@
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
 
-from tools.logging.logging_util import get_logger
-from services.abstract_service import *
 from config import *
+from services.abstract_service import *
+from tools.logging.logging_util import get_logger
 
 
 class ServiceCreator:
